♻️ Move zIndex constants and add convertDBStructureToNodes to features/reactflow/utils
