+++ .changeset/wicked-buses-think.md
@@ -1,0 +1,5 @@
+---
+"@liam-hq/erd-core": patch
+---
+
+♻️ Move zIndex constants and add convertDBStructureToNodes to features/reactflow/utils
